# authors_app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Authors API",
    "description": "API for managing authors, books, and companies",
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "auth",
      "description": "Operations related to users"
    },
    {
      "name": "book",
      "description": "Operations related to books"
    },
    {
      "name": "company",
      "description": "Operations related to companies"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "nagawaesther227@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                },
                "first_name": {
                  "type": "string",
                  "example": "Esther"
                },
                "last_name": {
                  "type": "string",
                  "example": "Nagawa"
                },
                "contact": {
                  "type": "integer",
                  "format": "int64",
                  "example": 712636785
                },
                "user_type": {
                  "type": "string",
                  "example": "author"
                },
                "biography": {
                  "type": "string",
                  "example": "Editor."
                }
              },
              "required": ["email", "password", "first_name", "last_name", "contact"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Missing email or password"
          },
          "409": {
            "description": "Email already exists"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login to obtain access token",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "nagawaesther227@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful. Access token generated."
          },
          "400": {
            "description": "Missing email or password"
          },
          "401": {
            "description": "Invalid email or password"
          }
        }
      }
    },
    "/auth/user/{user_id}": {
      "get": {
        "tags": ["auth"],
        "summary": "Get user details by ID",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "User details retrieved successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": ["auth"],
        "summary": "Update user details",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the user"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "nagawaesther227@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                },
                "first_name": {
                  "type": "string",
                  "example": "Esther"
                },
                "last_name": {
                  "type": "string",
                  "example": "Nagawa"
                },
                "contact": {
                  "type": "integer",
                  "format": "int64",
                  "example": 712636785
                },
                "user_type": {
                  "type": "string",
                  "example": "author"
                },
                "biography": {
                  "type": "string",
                  "example": "Editor."
                }
              },
              "required": ["email", "password", "first_name", "last_name", "contact"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details updated successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": ["auth"],
        "summary": "Delete user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the user"
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/auth/users": {
      "get": {
        "tags": ["auth"],
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "List of users retrieved successfully"
          }
        }
      }
    },
    "/book/register": {
      "post": {
        "tags": ["book"],
        "summary": "Create a new book",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Book Title"
                },
                "description": {
                  "type": "string",
                  "example": "Book description"
                },
                "price": {
                  "type": "integer",
                  "format": "int64",
                  "example": 20
                },
                "price_unit": {
                  "type": "string",
                  "example": "USD"
                },
                "pages": {
                  "type": "integer",
                  "format": "int64",
                  "example": 200
                },
                "publication_date": {
                  "type": "string",
                  "format": "date",
                  "example": "2024-04-15"
                },
                "isbn": {
                  "type": "string",
                  "example": "978-3-16-148410-0"
                },
                "genre": {
                  "type": "string",
                  "example": "Fiction"
                },
                "user_id": {
                  "type": "integer",
                  "format": "int64",
                  "example": 1
                },
                "company_id": {
                  "type": "integer",
                  "format": "int64",
                  "example": 1
                }
              },
              "required": ["title", "description", "price", "price_unit", "pages", "publication_date", "isbn", "genre", "user_id", "company_id"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Book created successfully"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/book/books": {
      "get": {
        "tags": ["book"],
        "summary": "Get all books",
        "responses": {
          "200": {
            "description": "List of books retrieved successfully"
          }
        }
      }
    },
    "/book/books/{book_id}": {
      "get": {
        "tags": ["book"],
        "summary": "Get book details by ID",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the book"
          }
        ],
        "responses": {
          "200": {
            "description": "Book details retrieved successfully"
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "put": {
        "tags": ["book"],
        "summary": "Update book details",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the book"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Book Title"
                },
                "description": {
                  "type": "string",
                  "example": "Book description"
                },
                "price": {
                  "type": "integer",
                  "format": "int64",
                  "example": 20
                },
                "price_unit": {
                  "type": "string",
                  "example": "USD"
                },
                "pages": {
                  "type": "integer",
                  "format": "int64",
                  "example": 200
                },
                "publication_date": {
                  "type": "string",
                  "format": "date",
                  "example": "2024-04-15"
                },
                "isbn": {
                  "type": "string",
                  "example": "978-3-16-148410-0"
                },
                "genre": {
                  "type": "string",
                  "example": "Fiction"
                },
                "user_id": {
                  "type": "integer",
                  "format": "int64",
                  "example": 1
                },
                "company_id": {
                  "type": "integer",
                  "format": "int64",
                  "example": 1
                }
              },
              "required": ["title", "description", "price", "price_unit", "pages", "publication_date", "isbn", "genre", "user_id", "company_id"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book updated successfully"
          },
          "404": {
            "description": "Book not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["book"],
        "summary": "Delete book",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the book"
          }
        ],
        "responses": {
          "204": {
            "description": "Book deleted successfully"
          },
          "404": {
            "description": "Book not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/company/register": {
      "post": {
        "tags": ["company"],
        "summary": "Create a new company",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Company Name"
                },
                "origin": {
                  "type": "string",
                  "example": "Company Origin"
                },
                "description": {
                  "type": "string",
                  "example": "Company Description"
                }
              },
              "required": ["name", "origin", "description"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Company created successfully"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/company/companies": {
      "get": {
        "tags": ["company"],
        "summary": "Get all companies",
        "responses": {
          "200": {
            "description": "List of companies retrieved successfully"
          }
        }
      }
    },
    "/company/companies/{company_id}": {
      "get": {
        "tags": ["company"],
        "summary": "Get company details by ID",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the company"
          }
        ],
        "responses": {
          "200": {
            "description": "Company details retrieved successfully"
          },
          "404": {
            "description": "Company not found"
          }
        }
      },
      "put": {
        "tags": ["company"],
        "summary": "Update company details",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the company"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Company Name"
                },
                "origin": {
                  "type": "string",
                  "example": "Company Origin"
                },
                "description": {
                  "type": "string",
                  "example": "Company Description"
                },
                "user_id": {
                  "type": "integer",
                  "format": "int64",
                  "example": 1
                }
              },
              "required": ["name", "origin", "description", "user_id"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company details updated successfully"
          },
          "404": {
            "description": "Company not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["company"],
        "summary": "Delete company",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the company"
          }
        ],
        "responses": {
          "204": {
            "description": "Company deleted successfully"
          },
          "404": {
            "description": "Company not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}